+++ CHANGELOG.md
@@ -11,16 +11,19 @@ All notable changes to this project will be documented in this file.
 - **CLI command `report category`** — support `--from` and `--to`.
 - **CLI command `report monthly`** — support `--from` and `--to`.
 - **CLI command `use`** — alias for the command `account use`.
-- **DTOs `CategoryItem::build`** — to resolve parent data.
+- **CLI command `report dashboard`** — support `--from` and `--to`.
+- **DTOs `CategoryItem::build`** — used to resolve parent data.
 - **DTOs `MonthlyReport::build`** — used by the new command `report monthly`.
 - **DTOs `CounterpartyStatsCollection::build`** — used by the new command `report counterpary`.
 - **DTOs `CategoryStatsCollection::build`** — used by the new command `report category`.
+- **DTOs `DashboardColection::build`** — used by the new command `report dashboard`.
 - **UI `view_montly`** — view for the new command `report monthly`.
 - **UI `view_counterparty_stats`** — stats view for the new command `report counterparty`.
 - **UI `view_category_stats`** — stats view for the new command `report category`.
+- **UI `view_dashboard_stats`** — stats view for the new command `report category`.
 
 ### Changed
-- **CLI command `report stats`** — rename to `report financial` to to clarify purpose.
+- **CLI command `report stats`** — rename to `report financial` to clarify purpose.
 -**UI** — Introduce unified UI style constants and helpers.
 -**UI** — account switch messaging to include the account name and id ("Switched to account: <name> (<id>)").
 - **UI `Top Expense`** — tie-breaker added to compare op_date so items with equal amounts are ordered deterministically.